1. CollisionMng 구현 - 함수 포인터를 통해 SceneMng가 가지고 있는 오브젝트 목록을 호출 할 수 있게 하는 기능 구현

+++ MathModule/Color.h
@@ -29,7 +29,7 @@ public:
 
 __forceinline constexpr Color Color::operator*(float InScalar) const {
 
-	return Color(R * InScalar, G * InScalar, B * InScalar, A * InScalar);
+	return Color((BYTE)(R * InScalar), (BYTE)(G * InScalar), (BYTE)(B * InScalar), (BYTE)(A * InScalar));
 }
